Use the package manager discovered from host's OS
This project was using only the `apt` module to install the
fail2ban package. But, it would be great to support multiple OS
distributions that has different package managers.

We can use fact variable `ansible_pkg_mgr`, discovered during the setup
task for each host, to know the name of the Ansible's package manager
module. By example: `yum` or `apt`. With this, We are able to invoke
the right module for each host dynamically using the `action` command.

```yml

- name: ensure fail2ban is installed¬
  action: "{{ ansible_pkg_mgr  }} name=fail2ban state=latest update_cache=yes"

```

Unfortunately, we can't continue to use the `apt_cache_valid_time`
and use only the common module properties: `name`, `state` and
`update_cache`.